requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchmetrics[image]>=1.0.0
tensorboard>=2.12.0
tqdm>=4.65.0
matplotlib>=3.7.0
pillow>=9.0.0
numpy>=1.23.0
ipython>=8.0.0